buildkite search: refactoring
