+++ misc/python/materialize/buildkite_insights/failure_search/search.py
@@ -24,13 +24,33 @@ from materialize.buildkite_insights.cache.cache_constants import (
     FETCH_MODE_NO,
     FETCH_MODE_YES,
 )
-from materialize.terminal import (
-    COLOR_CYAN,
-    COLOR_GREEN,
-    STYLE_BOLD,
-    with_formatting,
-    with_formattings,
+from materialize.buildkite_insights.failure_search.search_result_presentation import (
+    print_before_search_results,
+    print_match,
+    print_summary,
 )
+
+
+def search_build(build: Any, search_value: str, fetch_mode: str) -> int:
+    build_number = build["number"]
+    build_pipeline = build["pipeline"]["slug"]
+    build_state = build["state"]
+    web_url = build["web_url"]
+
+    is_completed_build_state = build_state in BUILDKITE_COMPLETED_BUILD_STATES
+
+    annotations = annotations_cache.get_or_query_annotations(
+        fetch_mode=fetch_mode,
+        pipeline_slug=build_pipeline,
+        build_number=build_number,
+        add_to_cache_if_not_present=is_completed_build_state,
+    )
+
+    matches_in_build = search_annotations(
+        build_number, build_pipeline, web_url, annotations, search_value
+    )
+
+    return matches_in_build
 
 
 def search_annotations(
@@ -63,89 +83,6 @@ def matches(annotation_text: str, search_value: str) -> bool:
     return search_value.lower() in annotation_text.lower()
 
 
-def highlight_match(annotation_text: str, search_value: str) -> str:
-    case_insensitive_pattern = re.compile(re.escape(search_value), re.IGNORECASE)
-    highlighted_match = with_formattings(search_value, [COLOR_GREEN, STYLE_BOLD])
-    return case_insensitive_pattern.sub(highlighted_match, annotation_text)
-
-
-def trim_match(annotation_text: str, search_value: str) -> str:
-    # We do not care about multiple occurrences within an annotation and focus on the first one.
-
-    original_annotation_text = annotation_text.strip()
-    annotation_text = original_annotation_text.lower()
-    search_value = search_value.lower()
-
-    max_chars_before_match = 300
-    max_chars_after_match = 300
-
-    match_begin_index = annotation_text.index(search_value)
-    match_end_index = match_begin_index + len(search_value)
-
-    # identify cut-off point before first match
-    if match_begin_index > max_chars_before_match:
-        cut_off_index_begin = annotation_text.find(
-            " ", match_begin_index - max_chars_before_match
-        )
-
-        if cut_off_index_begin == -1:
-            cut_off_index_begin = match_begin_index - max_chars_before_match
-    else:
-        cut_off_index_begin = 0
-
-    # identify cut-off point after first match
-    if len(annotation_text) - match_end_index > 300:
-        cut_off_index_end = annotation_text.rfind(
-            " ", match_end_index, match_end_index + max_chars_after_match
-        )
-
-        if cut_off_index_end == -1:
-            cut_off_index_end = match_end_index + max_chars_after_match
-    else:
-        cut_off_index_end = len(annotation_text)
-
-    cut_annotation_text = original_annotation_text[
-        cut_off_index_begin:cut_off_index_end
-    ]
-    cut_annotation_text = cut_annotation_text.strip()
-
-    if cut_off_index_begin > 0:
-        cut_annotation_text = f"[...] {cut_annotation_text}"
-
-    if cut_off_index_end != len(original_annotation_text):
-        cut_annotation_text = f"{cut_annotation_text} [...]"
-
-    return cut_annotation_text
-
-
-def print_match(
-    build_number: str,
-    build_pipeline: str,
-    web_url: str,
-    annotation_text: str,
-    search_value: str,
-) -> None:
-    match_snippet = trim_match(
-        annotation_text=annotation_text, search_value=search_value
-    )
-    match_snippet = highlight_match(
-        annotation_text=match_snippet,
-        search_value=search_value,
-    )
-
-    print(
-        with_formatting(
-            f"Match in build #{build_number} (pipeline={build_pipeline}):", STYLE_BOLD
-        )
-    )
-    print(f"URL: {with_formatting(web_url, COLOR_CYAN)}")
-    print("----------")
-    print(match_snippet)
-    print(
-        "-------------------------------------------------------------------------------------"
-    )
-
-
 def main(
     pipeline_slug: str,
     fetch_mode: str,
@@ -175,41 +112,15 @@ def main(
             build_states=build_states,
         )
 
-    print()
-    print(
-        "-------------------------------------------------------------------------------------"
-    )
+    print_before_search_results()
 
     count_matches = 0
 
     for build in builds_data:
-        build_number = build["number"]
-        build_pipeline = build["pipeline"]["slug"]
-        build_state = build["state"]
-        web_url = build["web_url"]
-
-        is_completed_build_state = build_state in BUILDKITE_COMPLETED_BUILD_STATES
-
-        annotations = annotations_cache.get_or_query_annotations(
-            fetch_mode=fetch_mode,
-            pipeline_slug=build_pipeline,
-            build_number=build_number,
-            add_to_cache_if_not_present=is_completed_build_state,
-        )
-
-        matches_in_build = search_annotations(
-            build_number, build_pipeline, web_url, annotations, search_value
-        )
+        matches_in_build = search_build(build, search_value, fetch_mode=fetch_mode)
         count_matches = count_matches + matches_in_build
 
-    if len(builds_data) == 0:
-        print("Found no builds!")
-    else:
-        most_recent_build_number = builds_data[0]["number"]
-        print(
-            f"{count_matches} match(es) in {len(builds_data)} searched builds of pipeline '{pipeline_slug}'. "
-            f"The most recent considered build was #{most_recent_build_number}."
-        )
+    print_summary(pipeline_slug, builds_data, count_matches)
 
 
 if __name__ == "__main__":
